+++ src/components/FloatingAction/styles.ts
@@ -1,6 +1,7 @@
 import styled from "styled-components/native";
 import { BorderlessButton } from "react-native-gesture-handler";
 import Animated from "react-native-reanimated";
+import { RFValue } from "react-native-responsive-fontsize";
 
 const AnimatedBorderlessButton = Animated.createAnimatedComponent(BorderlessButton);
 
@@ -8,8 +9,8 @@ export const Container = styled(AnimatedBorderlessButton)`
   align-items: center;
   justify-content: center;
 
-  width: 48px;
-  height: 48px;
-  border-radius: 24px;
+  width: ${RFValue(48)}px;
+  height: ${RFValue(48)}px;
+  border-radius: ${RFValue(24)}px;
   background: ${({ theme }) => theme.colors.secundary};
 `;