+++ src/components/MainFooter/index.tsx
@@ -1,6 +1,7 @@
 import { FC, useEffect, useState } from "react";
 import { useTheme } from "styled-components/native";
 import { StyleSheet } from "react-native";
+import { useNavigation } from "@react-navigation/native";
 
 import { useAnimation } from "../../hooks/useAnimation";
 import { useGlobal } from "../../hooks/useGlobal";
@@ -13,14 +14,17 @@ import {
   ScreenBadgeText
 } from "./styles";
 
+type ChangeScreenProps = "Tasks" | "Trash" | "Favorit";
+
 export const MainFooter: FC = () => {
   const { 
     changeBackgroundColorNavigation, 
     onChangeCurrentScreen,
     currentScreen
   } = useGlobal();
-  const { mainFooterAnimation } = useAnimation();
+  const { mainFooterAnimation, onHideMainFooter } = useAnimation();
   const { colors } = useTheme();
+  const { isFocused, goBack, navigate} = useNavigation();
 
   const [countRemovedTask, setCountRemovedTask] = useState(0);
 
@@ -35,11 +39,32 @@ export const MainFooter: FC = () => {
     }
   });
 
-  useEffect(() => {
-    setTimeout(() => {
-      setCountRemovedTask(10);
-    }, 3000);
-  }, []);
+  function onChangeScheen(screen: ChangeScreenProps) {
+    if (currentScreen === screen) {
+      return;
+    }
+
+    if (currentScreen !== screen && screen !== "Tasks") {
+      onChangeCurrentScreen(screen);
+
+      onHideMainFooter(() => {
+        navigate(screen);
+      });
+    }
+
+    if (screen === "Tasks") {
+      onChangeCurrentScreen(screen);
+
+      onHideMainFooter();
+      goBack();
+    }
+  }
+
+  // useEffect(() => {
+  //   setTimeout(() => {
+  //     setCountRemovedTask(10);
+  //   }, 3000);
+  // }, []);
 
   useEffect(() => {
     changeBackgroundColorNavigation(colors.secundary);
@@ -48,19 +73,19 @@ export const MainFooter: FC = () => {
   return (
     <Container style={[mainFooterAnimation, styles.container]}>
       <ScreenButton 
-        isActive={currentScreen === "Tasks" && true}
-        onPress={() => onChangeCurrentScreen("Tasks")}
+        isActive={currentScreen === "Tasks" && isFocused()}
+        onPress={() => onChangeScheen("Tasks")}
       >
         <ScrennAnimatedIcon 
-          isActive={currentScreen === "Tasks" && true}
+          isActive={currentScreen === "Tasks" && isFocused()}
           name="grid" 
           size={currentScreen === "Tasks" ? 24 : 20}
         />
       </ScreenButton>
 
       <ScreenButton 
-        isActive={currentScreen === "Trash" && true}
-        onPress={() => onChangeCurrentScreen("Trash")}
+        isActive={currentScreen === "Trash" && isFocused()}
+        onPress={() => onChangeScheen("Trash")}
       >
         {countRemovedTask > 0 && (
           <ScreenBadge>
@@ -68,18 +93,18 @@ export const MainFooter: FC = () => {
           </ScreenBadge>
         )}
         <ScrennAnimatedIcon 
-          isActive={currentScreen === "Trash" && true}
+          isActive={currentScreen === "Trash" && isFocused()}
           name={countRemovedTask > 0 ? "trash-2" : "trash"} 
           size={currentScreen === "Trash" ? 24 : 20}
         />
       </ScreenButton>
 
       <ScreenButton 
-        isActive={currentScreen === "Favorit" && true}
-        onPress={() => onChangeCurrentScreen("Favorit")}
+        isActive={currentScreen === "Favorit" && isFocused()}
+        onPress={() => onChangeScheen("Favorit")}
       >
         <ScrennAnimatedIcon 
-          isActive={currentScreen === "Favorit" && true}
+          isActive={currentScreen === "Favorit" && isFocused()}
           name="star" 
           size={currentScreen === "Favorit" ? 24 : 20}
         />