+++ src/routers/index.tsx
@@ -1,9 +1,9 @@
 import React, { FC, useEffect } from "react";
 import { NavigationContainer } from "@react-navigation/native";
-import { createStackNavigator } from "@react-navigation/stack";
+import { createStackNavigator, CardStyleInterpolators } from "@react-navigation/stack";
 import * as ScreenOrientation from 'expo-screen-orientation';
 
-import { Tasks } from "../screens/Tasks";
+import { Tasks, Trash } from "../screens";
 
 const Stack = createStackNavigator();
 
@@ -24,11 +24,18 @@ export const Router: FC = () => {
       <Stack.Navigator
         screenOptions={{
           headerShown: false,
+          cardStyleInterpolator: CardStyleInterpolators.forHorizontalIOS,
+          cardShadowEnabled: false
         }}
       >
         <Stack.Screen 
           name="Tasks" 
           component={Tasks} 
+        />
+
+        <Stack.Screen 
+          name="Trash" 
+          component={Trash} 
         />
       </Stack.Navigator>
     </NavigationContainer>