+++ src/screens/index.ts
@@ -1,0 +1,4 @@
+import { Tasks } from "./Tasks";
+import { Trash } from "./Trash";
+
+export { Tasks, Trash };